backend/requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
python-multipart>=0.0.9,<1.0.0
pydantic>=2.8.0,<3.0.0
httpx>=0.27.0,<1.0.0
azure-cosmos>=4.7.0,<5.0.0
azure-storage-blob>=12.21.0,<13.0.0
azure-communication-email>=1.0.0,<2.0.0
azure-identity>=1.17.1,<2.0.0
azure-keyvault-secrets>=4.8.0,<5.0.0
azure-keyvault-keys>=4.10.0,<5.0.0
azure-monitor-opentelemetry>=1.6.0,<2.0.0
cryptography>=43.0.0,<46.0.0
opentelemetry-instrumentation-fastapi>=0.48b0,<1.0.0
